reduce block number to uint128; reduces gas used by 50k-150k per operation

+++ contracts/IFAllocationMaster.sol
@@ -23,7 +23,7 @@ contract IFAllocationMaster is Ownable {
     // A checkpoint for marking stake info at a given block
     struct UserCheckpoint {
         // block number of checkpoint
-        uint256 blockNumber;
+        uint128 blockNumber;
         // amount staked at checkpoint
         uint104 staked;
         // amount of stake weight at checkpoint
@@ -35,7 +35,7 @@ contract IFAllocationMaster is Ownable {
     // A checkpoint for marking stake info at a given block
     struct TrackCheckpoint {
         // block number of checkpoint
-        uint256 blockNumber;
+        uint128 blockNumber;
         // amount staked at checkpoint
         uint104 totalStaked;
         // amount of stake weight at checkpoint
@@ -67,7 +67,7 @@ contract IFAllocationMaster is Ownable {
     // INFO FOR FACTORING IN ROLLOVERS
 
     // the number of checkpoints of a track -- (track, finished sale count) => block number
-    mapping(uint24 => mapping(uint24 => uint256))
+    mapping(uint24 => mapping(uint24 => uint128))
         public trackFinishedSaleBlocks;
 
     // stake weight each user actively rolls over for a given track and a finished sale count
@@ -106,8 +106,8 @@ contract IFAllocationMaster is Ownable {
     event AddTrack(string indexed name, address indexed token);
     event DisableTrack(uint24 indexed trackId);
     event BumpSaleCounter(uint24 indexed trackId, uint32 newCount);
-    event AddUserCheckpoint(uint256 blockNumber, uint24 indexed trackId);
-    event AddTrackCheckpoint(uint256 blockNumber, uint24 indexed trackId);
+    event AddUserCheckpoint(uint128 blockNumber, uint24 indexed trackId);
+    event AddTrackCheckpoint(uint128 blockNumber, uint24 indexed trackId);
     event Stake(address indexed user, uint24 indexed trackId, uint104 amount);
     event Unstake(address indexed user, uint24 indexed trackId, uint104 amount);
 
@@ -162,7 +162,9 @@ contract IFAllocationMaster is Ownable {
                 .numFinishedSales;
 
         // update map that tracks block numbers of finished sales
-        trackFinishedSaleBlocks[trackId][nFinishedSales] = block.number;
+        trackFinishedSaleBlocks[trackId][nFinishedSales] = uint128(
+            block.number
+        );
 
         // add a new checkpoint with counter incremented by 1
         addTrackCheckpoint(trackId, 0, false, false, true);
@@ -209,7 +211,7 @@ contract IFAllocationMaster is Ownable {
     function getClosestUserCheckpoint(
         uint24 trackId,
         address user,
-        uint256 blockNumber
+        uint128 blockNumber
     ) private view returns (UserCheckpoint memory cp) {
         // get total checkpoint count for user
         uint32 nCheckpoints = userCheckpointCounts[trackId][user];
@@ -263,7 +265,7 @@ contract IFAllocationMaster is Ownable {
     function getUserStakeWeight(
         uint24 trackId,
         address user,
-        uint256 blockNumber
+        uint128 blockNumber
     ) public view returns (uint104) {
         require(blockNumber <= block.number, 'block # too high');
 
@@ -299,7 +301,7 @@ contract IFAllocationMaster is Ownable {
         if (numFinishedSalesDelta == 0) {
             // calculate normally without rollover decay
 
-            uint256 elapsedBlocks =
+            uint128 elapsedBlocks =
                 blockNumber - closestUserCheckpoint.blockNumber;
 
             stakeWeight =
@@ -316,13 +318,13 @@ contract IFAllocationMaster is Ownable {
             // starting stakeweight
             stakeWeight = closestUserCheckpoint.stakeWeight;
             // current block for iteration
-            uint256 currBlock = closestUserCheckpoint.blockNumber;
+            uint128 currBlock = closestUserCheckpoint.blockNumber;
 
             // for each finished sale in between, get stake weight of that period
             // and perform weighted sum
             for (uint24 i = 0; i < numFinishedSalesDelta; i++) {
                 // get number of blocks passed at the current sale number
-                uint256 elapsedBlocks =
+                uint128 elapsedBlocks =
                     trackFinishedSaleBlocks[trackId][
                         closestUserCheckpoint.numFinishedSales + i
                     ] - currBlock;
@@ -358,7 +360,7 @@ contract IFAllocationMaster is Ownable {
             }
 
             // add any remaining accrued stake weight at current finished sale count
-            uint256 remainingElapsed =
+            uint128 remainingElapsed =
                 blockNumber -
                     trackFinishedSaleBlocks[trackId][
                         closestTrackCheckpoint.numFinishedSales - 1
@@ -375,7 +377,7 @@ contract IFAllocationMaster is Ownable {
     }
 
     // get closest PRECEDING track checkpoint
-    function getClosestTrackCheckpoint(uint24 trackId, uint256 blockNumber)
+    function getClosestTrackCheckpoint(uint24 trackId, uint128 blockNumber)
         private
         view
         returns (TrackCheckpoint memory cp)
@@ -429,10 +431,10 @@ contract IFAllocationMaster is Ownable {
 
     // gets total stake weight within a track at a particular block number
     // logic extended from Compound COMP token `getPriorVotes` function
-    function getTotalStakeWeight(uint24 trackId, uint256 blockNumber)
+    function getTotalStakeWeight(uint24 trackId, uint128 blockNumber)
         public
         view
-        returns (uint256)
+        returns (uint104)
     {
         require(blockNumber <= block.number, 'block # too high');
 
@@ -441,15 +443,15 @@ contract IFAllocationMaster is Ownable {
             getClosestTrackCheckpoint(trackId, blockNumber);
 
         // calculate blocks elapsed since checkpoint
-        uint256 additionalBlocks =
+        uint128 additionalBlocks =
             (blockNumber - closestCheckpoint.blockNumber);
 
         // get track info
         TrackInfo storage trackInfo = tracks[trackId];
 
         // calculate marginal accrued stake weight
-        uint256 marginalAccruedStakeWeight =
-            (additionalBlocks *
+        uint104 marginalAccruedStakeWeight =
+            (uint104(additionalBlocks) *
                 trackInfo.weightAccrualRate *
                 closestCheckpoint.totalStaked) / 10**18;
 
@@ -495,7 +497,7 @@ contract IFAllocationMaster is Ownable {
 
             // add a first checkpoint for this user on this track
             userCheckpoints[trackId][_msgSender()][0] = UserCheckpoint({
-                blockNumber: block.number,
+                blockNumber: uint128(block.number),
                 staked: amount,
                 stakeWeight: 0,
                 numFinishedSales: trackCp.numFinishedSales
@@ -509,14 +511,14 @@ contract IFAllocationMaster is Ownable {
             userCheckpoints[trackId][_msgSender()][
                 nCheckpointsUser
             ] = UserCheckpoint({
-                blockNumber: block.number,
+                blockNumber: uint128(block.number),
                 staked: addElseSub
                     ? prev.staked + amount
                     : prev.staked - amount,
                 stakeWeight: getUserStakeWeight(
                     trackId,
                     _msgSender(),
-                    block.number
+                    uint128(block.number)
                 ),
                 numFinishedSales: trackCp.numFinishedSales
             });
@@ -541,7 +543,7 @@ contract IFAllocationMaster is Ownable {
         userCheckpointCounts[trackId][_msgSender()] = nCheckpointsUser + 1;
 
         // emit
-        emit AddUserCheckpoint(block.number, trackId);
+        emit AddUserCheckpoint(uint128(block.number), trackId);
     }
 
     function addTrackCheckpoint(
@@ -561,7 +563,7 @@ contract IFAllocationMaster is Ownable {
         if (nCheckpoints == 0) {
             // add a first checkpoint for this track
             trackCheckpoints[trackId][0] = TrackCheckpoint({
-                blockNumber: block.number,
+                blockNumber: uint128(block.number),
                 totalStaked: amount,
                 totalStakeWeight: 0,
                 disabled: disabled,
@@ -580,7 +582,8 @@ contract IFAllocationMaster is Ownable {
                 trackCheckpoints[trackId][nCheckpoints - 1];
 
             // calculate blocks elapsed since checkpoint
-            uint256 additionalBlocks = (block.number - prev.blockNumber);
+            uint128 additionalBlocks =
+                (uint128(block.number) - prev.blockNumber);
 
             // calculate marginal accrued stake weight
             uint104 marginalAccruedStakeWeight =
@@ -634,7 +637,7 @@ contract IFAllocationMaster is Ownable {
                 // if previous checkpoint was disabled, stakeweight cannot increase
                 // and new checkpoint must also be disabled
                 trackCheckpoints[trackId][nCheckpoints] = TrackCheckpoint({
-                    blockNumber: block.number,
+                    blockNumber: uint128(block.number),
                     totalStaked: prev.totalStaked - amount,
                     totalStakeWeight: prev.totalStakeWeight,
                     disabled: true,
@@ -643,7 +646,7 @@ contract IFAllocationMaster is Ownable {
                 });
             } else {
                 trackCheckpoints[trackId][nCheckpoints] = TrackCheckpoint({
-                    blockNumber: block.number,
+                    blockNumber: uint128(block.number),
                     totalStaked: addElseSub
                         ? prev.totalStaked + amount
                         : prev.totalStaked - amount,
@@ -669,7 +672,7 @@ contract IFAllocationMaster is Ownable {
         trackCheckpointCounts[trackId]++;
 
         // emit
-        emit AddTrackCheckpoint(block.number, trackId);
+        emit AddTrackCheckpoint(uint128(block.number), trackId);
     }
 
     // stake
